fix warning

+++ PhotoBoard/Classes/Presenters/PBWeChatActivity.m
@@ -21,7 +21,7 @@
     return UIActivityCategoryShare;
 }
 
-- (NSInteger)scene {
+- (int)scene {
     return WXSceneSession;
 }
